+++ examples/complete/main.tf
@@ -36,65 +36,82 @@ module "complete_eks_cluster" {
     },
   ]
   managed_node_groups = {
-    managed0 = {
-      create       = true
-      subnet_ids   = local.private_subnets
-      desired_size = 3
-      max_size     = 3
-      min_size     = 1
-      tags         = var.tags
+    # managed0 = {
+    #   create       = true
+    #   subnet_ids   = local.private_subnets
+    #   desired_size = 3
+    #   max_size     = 3
+    #   min_size     = 1
+    #   tags         = local.tags
 
-      # launch template
-      create_custom_launch_template = true
-      launch_template_description   = "EKS managed node group launch template"
-      ebs_optimized                 = true
-      install_ssm_agent             = true
-      block_device_mappings = [
-        {
-          # Root volume
-          device_name = "/dev/xvda"
-          no_device   = 0
-          ebs = {
-            delete_on_termination = true
-            volume_size           = 30
-            volume_type           = "gp3"
-          }
-        },
-        {
-          device_name = "/dev/sda1"
-          no_device   = 1
-          ebs = {
-            delete_on_termination = true
-            volume_size           = 30
-            volume_type           = "gp2"
-          }
-        }
-      ]
+    #   # launch template
+    #   create_custom_launch_template = true
+    #   launch_template_description   = "EKS managed node group launch template"
+    #   ebs_optimized                 = true
+    #   install_ssm_agent             = true
+    #   block_device_mappings = [
+    #     {
+    #       # Root volume
+    #       device_name = "/dev/xvda"
+    #       no_device   = 0
+    #       ebs = {
+    #         delete_on_termination = true
+    #         volume_size           = 30
+    #         volume_type           = "gp3"
+    #       }
+    #     },
+    #     {
+    #       device_name = "/dev/sda1"
+    #       no_device   = 1
+    #       ebs = {
+    #         delete_on_termination = true
+    #         volume_size           = 30
+    #         volume_type           = "gp2"
+    #       }
+    #     }
+    #   ]
 
-      metadata_options = {
-        http_endpoint               = "enabled"
-        http_tokens                 = "required"
-        http_put_response_hop_limit = 2
-        instance_metadata_tags      = "disabled"
-      }
+    #   metadata_options = {
+    #     http_endpoint               = "enabled"
+    #     http_tokens                 = "required"
+    #     http_put_response_hop_limit = 2
+    #     instance_metadata_tags      = "disabled"
+    #   }
 
-      tag_specifications = [
-        {
-          resource_type = "volume"
-          tags          = local.tags
-        },
-        {
-          resource_type = "instance"
-          tags          = local.tags
-        }
-      ]
+    #   tag_specifications = [
+    #     {
+    #       resource_type = "volume"
+    #       tags          = local.tags
+    #     },
+    #     {
+    #       resource_type = "instance"
+    #       tags          = local.tags
+    #     }
+    #   ]
 
-    }
+    # }
     managed1 = {
       create        = true
       subnet_ids    = local.private_subnets
       capacity_type = "SPOT"
-      tags          = var.tags
+      taints = {
+        dedicated = {
+          "key"    = "dedicated"
+          "value"  = "true"
+          "effect" = "NO_SCHEDULE"
+        }
+        example1 = {
+          "key"    = "example1"
+          "value"  = "true"
+          "effect" = "NO_EXECUTE"
+        }
+        example1 = {
+          "key"    = "example2"
+          "value"  = "true"
+          "effect" = "PREFER_NO_SCHEDULE"
+        }
+      }
+      tags = local.tags
     }
   }
 
@@ -109,7 +126,7 @@ module "complete_eks_cluster" {
         }
       ]
       subnet_ids = local.private_subnets
-      tags       = var.tags
+      tags       = local.tags
     }
     fargate1 = {
       selector = [
@@ -121,8 +138,8 @@ module "complete_eks_cluster" {
         }
       ]
       subnet_ids = local.private_subnets
-      tags       = var.tags
+      tags       = local.tags
     }
   }
-  tags = var.tags
+  tags = local.tags
 }